fix(suite, suite-data): adjust backport-en script to changes in messages.ts
