+++ packages/suite-data/src/translations/backport-en.ts
@@ -38,7 +38,9 @@ fs.writeFileSync(
     `
 import { defineMessages } from 'react-intl';
 
-export default defineMessages(${JSON.stringify(messages, null, 2).replace(/"([^"]+)":/g, '$1:')})
+import { defineMessagesWithTypeCheck } from './utils';
+
+const messages = defineMessagesWithTypeCheck(${JSON.stringify(messages, null, 2).replace(/"([^"]+)":/g, '$1:')})
 
 `,
 );
